# backend-go/internal/handler/dashboard.go
package handler

import (
	"errors"
	"io"
	"strconv"

	"github.com/BenedictKing/new_api_tools/internal/cache"
	"github.com/BenedictKing/new_api_tools/internal/logger"
	"github.com/BenedictKing/new_api_tools/internal/service"
	"github.com/gin-gonic/gin"
	"go.uber.org/zap"
)

var dashboardService = service.NewDashboardService()
var ipDistributionService = service.NewIPDistributionService()

// GetDashboardOverview 获取系统概览
// @Summary     获取系统概览
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Success     200  {object}  Response{data=object}
// @Router      /dashboard/overview [get]
func GetDashboardOverview(c *gin.Context) {
	data, err := dashboardService.GetOverview()
	if err != nil {
		logger.Error("获取系统概览失败", zap.Error(err))
		Error(c, 500, "获取系统概览失败")
		return
	}

	Success(c, data)
}

// GetDashboardUsage 获取使用统计
// @Summary     获取使用统计
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Param       period  query     string  false  "时间周期 (today/7d/30d)"  default(today)
// @Success     200     {object}  Response{data=object}
// @Router      /dashboard/usage [get]
func GetDashboardUsage(c *gin.Context) {
	period := c.DefaultQuery("period", "today")

	data, err := dashboardService.GetUsage(period)
	if err != nil {
		logger.Error("获取使用统计失败", zap.Error(err))
		Error(c, 500, "获取使用统计失败")
		return
	}

	Success(c, data)
}

// GetDashboardModels 获取模型使用统计
// @Summary     获取模型使用统计
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Param       period  query     string  false  "时间周期"  default(24h)
// @Param       limit   query     int     false  "返回数量"  default(10)
// @Success     200     {object}  Response{data=object}
// @Router      /dashboard/models [get]
func GetDashboardModels(c *gin.Context) {
	period := c.DefaultQuery("period", "24h")
	limit, _ := strconv.Atoi(c.DefaultQuery("limit", "10"))

	data, err := dashboardService.GetModelUsage(period, limit)
	if err != nil {
		logger.Error("获取模型统计失败", zap.Error(err))
		Error(c, 500, "获取模型统计失败")
		return
	}

	Success(c, data)
}

// GetDailyTrends 获取每日趋势
// @Summary     获取每日趋势
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Param       days  query     int  false  "天数"  default(30)
// @Success     200   {object}  Response{data=object}
// @Router      /dashboard/trends/daily [get]
func GetDailyTrends(c *gin.Context) {
	days, _ := strconv.Atoi(c.DefaultQuery("days", "30"))

	data, err := dashboardService.GetDailyTrends(days)
	if err != nil {
		logger.Error("获取每日趋势失败", zap.Error(err))
		Error(c, 500, "获取每日趋势失败")
		return
	}

	Success(c, data)
}

// GetHourlyTrends 获取每小时趋势
// @Summary     获取每小时趋势
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Param       hours  query     int  false  "小时数 (1-168)"  default(24)
// @Success     200    {object}  Response{data=object}
// @Router      /dashboard/trends/hourly [get]
func GetHourlyTrends(c *gin.Context) {
	hours, _ := strconv.Atoi(c.DefaultQuery("hours", "24"))
	if hours <= 0 || hours > 168 { // 最多 7 天
		hours = 24
	}

	data, err := dashboardService.GetHourlyTrends(hours)
	if err != nil {
		logger.Error("获取每小时趋势失败", zap.Error(err))
		Error(c, 500, "获取每小时趋势失败")
		return
	}

	Success(c, data)
}

// GetTopUsers 获取用户排行
// @Summary     获取用户排行榜
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Param       period    query     string  false  "时间周期"         default(24h)
// @Param       limit     query     int     false  "返回数量 (1-100)"  default(10)
// @Param       order_by  query     string  false  "排序字段 (requests/quota)"  default(requests)
// @Success     200       {object}  Response{data=object}
// @Router      /dashboard/top-users [get]
func GetTopUsers(c *gin.Context) {
	period := c.DefaultQuery("period", "24h")
	limit, _ := strconv.Atoi(c.DefaultQuery("limit", "10"))
	if limit <= 0 || limit > 100 {
		limit = 10
	}
	orderBy := c.DefaultQuery("order_by", "requests")

	data, err := dashboardService.GetTopUsers(period, limit, orderBy)
	if err != nil {
		logger.Error("获取用户排行失败", zap.Error(err))
		Error(c, 500, "获取用户排行失败")
		return
	}

	Success(c, data)
}

// GetChannelStatus 获取渠道状态
// @Summary     获取渠道状态
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Success     200  {object}  Response{data=object}
// @Router      /dashboard/channels [get]
func GetChannelStatus(c *gin.Context) {
	data, err := dashboardService.GetChannelStatus()
	if err != nil {
		logger.Error("获取渠道状态失败", zap.Error(err))
		Error(c, 500, "获取渠道状态失败")
		return
	}

	Success(c, data)
}

// GetIPDistribution 获取 IP 分布
// @Summary     获取 IP 地理分布
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Param       window  query     string  false  "时间窗口"  default(24h)
// @Success     200     {object}  Response{data=object}
// @Router      /dashboard/ip-distribution [get]
func GetIPDistribution(c *gin.Context) {
	window := c.DefaultQuery("window", "24h")

	data, err := ipDistributionService.GetDistribution(window)
	if err != nil {
		logger.Error("获取 IP 分布失败", zap.Error(err))
		Error(c, 500, "获取 IP 分布失败")
		return
	}

	Success(c, data)
}

// GetSystemInfo 获取系统信息
// @Summary     获取系统信息
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Success     200  {object}  Response{data=object}
// @Router      /dashboard/system-info [get]
func GetSystemInfo(c *gin.Context) {
	data, err := dashboardService.GetSystemInfo()
	if err != nil {
		logger.Error("获取系统信息失败", zap.Error(err))
		Error(c, 500, "获取系统信息失败")
		return
	}
	Success(c, data)
}

// InvalidateCache 清除缓存
// @Summary     清除仪表盘缓存
// @Tags        仪表盘
// @Accept      json
// @Produce     json
// @Security    BearerAuth
// @Param       body  body      object  false  "指定要清除的缓存 keys，空则清除全部"
// @Success     200   {object}  Response{data=object}
// @Router      /dashboard/cache/invalidate [post]
func InvalidateCache(c *gin.Context) {
	var req struct {
		Keys []string `json:"keys"`
	}

	// 尝试绑定 JSON，忽略 EOF 错误（空 body 的情况）
	if err := c.ShouldBindJSON(&req); err != nil {
		// io.EOF 表示空 body，这是允许的（清除所有缓存）
		if !errors.Is(err, io.EOF) {
			Error(c, 400, "请求格式错误: "+err.Error())
			return
		}
	}

	var totalCleared int64
	if len(req.Keys) == 0 {
		// 如果没有指定 keys，清除所有 dashboard 缓存
		deleted, err := cache.DeletePattern(cache.CacheKey("dashboard", "*"))
		if err != nil {
			logger.Error("清除 dashboard 缓存失败", zap.Error(err))
		}
		totalCleared = deleted
	} else {
		// 根据指定的 keys 清除对应缓存
		for _, key := range req.Keys {
			pattern := cache.CacheKey("dashboard", key, "*")
			deleted, err := cache.DeletePattern(pattern)
			if err != nil {
				logger.Warn("清除缓存失败", zap.String("key", key), zap.Error(err))
			}
			totalCleared += deleted
		}
	}

	Success(c, gin.H{"message": "缓存已清除", "cleared": totalCleared})
}

// GetRefreshEstimate 获取刷新时间估算
// @Summary     获取缓存刷新时间估算
// @Tags        仪表盘
// @Produce     json
// @Security    BearerAuth
// @Param       period  query     string  false  "时间周期"  default(24h)
// @Success     200     {object}  Response{data=object}
// @Router      /dashboard/refresh-estimate [get]
func GetRefreshEstimate(c *gin.Context) {
	period := c.DefaultQuery("period", "24h")

	systemService := service.NewSystemService()
	data, err := systemService.GetRefreshEstimate(period)
	if err != nil {
		logger.Error("获取刷新时间估算失败", zap.Error(err))
		// 返回默认值
		Success(c, gin.H{
			"estimated_seconds": 30,
			"last_refresh":      nil,
		})
		return
	}

	Success(c, data)
}

// ==================== Top-Ups ====================
func GetTopUps(c *gin.Context)          { GetTopUpsHandler(c) }
func GetTopUpStatistics(c *gin.Context) { GetTopUpStatisticsHandler(c) }
func GetPaymentMethods(c *gin.Context)  { GetPaymentMethodsHandler(c) }
func RefundTopUp(c *gin.Context)        { RefundTopUpHandler(c) }

// ==================== Redemptions ====================
func GenerateRedemptions(c *gin.Context)     { GenerateRedemptionsHandler(c) }
func GetRedemptions(c *gin.Context)          { GetRedemptionsHandler(c) }
func GetRedemptionStatistics(c *gin.Context) { GetRedemptionStatisticsHandler(c) }
func DeleteRedemption(c *gin.Context)        { DeleteRedemptionHandler(c) }
func BatchDeleteRedemptions(c *gin.Context)  { BatchDeleteRedemptionsHandler(c) }

// ==================== Users ====================
func GetUsers(c *gin.Context)                 { GetUsersHandler(c) }
func GetUserStats(c *gin.Context)             { GetUserStatsHandler(c) }
func GetBannedUsers(c *gin.Context)           { GetBannedUsersHandler(c) }
func GetSoftDeletedUsersCount(c *gin.Context) { GetSoftDeletedUsersCountHandler(c) }
func PurgeSoftDeletedUsers(c *gin.Context)    { PurgeSoftDeletedUsersHandler(c) }
func DeleteUser(c *gin.Context)               { DeleteUserHandler(c) }
func BatchDeleteUsers(c *gin.Context)         { BatchDeleteUsersHandler(c) }
func BanUser(c *gin.Context)                  { BanUserHandler(c) }
func UnbanUser(c *gin.Context)                { UnbanUserHandler(c) }
func DisableToken(c *gin.Context)             { DisableTokenHandler(c) }
func GetInvitedUsers(c *gin.Context)          { GetInvitedUsersHandler(c) }

// ==================== Risk Monitoring ====================
func GetLeaderboards(c *gin.Context)        { GetLeaderboardsHandler(c) }
func GetUserRiskAnalysis(c *gin.Context)    { GetUserRiskAnalysisHandler(c) }
func GetBanRecords(c *gin.Context)          { GetBanRecordsHandler(c) }
func GetTokenRotation(c *gin.Context)       { GetTokenRotationHandler(c) }
func GetAffiliatedAccounts(c *gin.Context)  { GetAffiliatedAccountsHandler(c) }
func GetSameIPRegistrations(c *gin.Context) { GetSameIPRegistrationsHandler(c) }

// ==================== IP Monitoring ====================
func GetIPStats(c *gin.Context)       { GetIPStatsHandler(c) }
func GetSharedIPs(c *gin.Context)     { GetSharedIPsHandler(c) }
func GetMultiIPTokens(c *gin.Context) { GetMultiIPTokensHandler(c) }
func GetMultiIPUsers(c *gin.Context)  { GetMultiIPUsersHandler(c) }
func EnableAllIPRecording(c *gin.Context) {
	ipService := service.NewIPService()
	result, err := ipService.EnableAllIPRecording()
	if err != nil {
		logger.Error("批量开启 IP 记录失败", zap.Error(err))
		Error(c, 500, "批量开启 IP 记录失败: "+err.Error())
		return
	}

	c.JSON(200, Response{
		Success: true,
		Message: "已更新 " + strconv.FormatInt(result.UpdatedCount, 10) + " 个用户，跳过 " + strconv.FormatInt(result.SkippedCount, 10) + " 个已开启的用户",
		Data:    result,
	})
}
func GetIPGeo(c *gin.Context)      { GetIPGeoHandler(c) }
func BatchGetIPGeo(c *gin.Context) { BatchGetIPGeoHandler(c) }

// ==================== AI Ban ====================
func GetAIBanConfig(c *gin.Context)     { GetAIBanConfigHandler(c) }
func UpdateAIBanConfig(c *gin.Context)  { UpdateAIBanConfigHandler(c) }
func TestAIModel(c *gin.Context)        { TestAIModelHandler(c) }
func GetSuspiciousUsers(c *gin.Context) { GetSuspiciousUsersHandler(c) }
func AssessUserRisk(c *gin.Context)     { AssessUserRiskHandler(c) }
func ScanUsers(c *gin.Context)          { ScanUsersHandler(c) }
func GetWhitelist(c *gin.Context)       { GetWhitelistHandler(c) }
func AddToWhitelist(c *gin.Context)     { AddToWhitelistHandler(c) }

// ==================== Analytics ====================
func GetAnalyticsState(c *gin.Context)     { GetAnalyticsStateHandler(c) }
func ProcessLogs(c *gin.Context)           { ProcessLogsHandler(c) }
func GetUserRequestRanking(c *gin.Context) { GetUserRequestRankingHandler(c) }
func GetUserQuotaRanking(c *gin.Context)   { GetUserQuotaRankingHandler(c) }
func GetModelStats(c *gin.Context)         { GetModelStatsHandler(c) }
func GetAnalyticsSummary(c *gin.Context)   { GetAnalyticsSummaryHandler(c) }
func ResetAnalytics(c *gin.Context)        { ResetAnalyticsHandler(c) }

// ==================== Model Status ====================
func GetAvailableModels(c *gin.Context)   { GetAvailableModelsHandler(c) }
func GetModelStatus(c *gin.Context)       { GetModelStatusHandler(c) }
func BatchGetModelStatus(c *gin.Context)  { BatchGetModelStatusHandler(c) }
func GetSelectedModels(c *gin.Context)    { GetSelectedModelsHandler(c) }
func UpdateSelectedModels(c *gin.Context) { UpdateSelectedModelsHandler(c) }
func GetTimeWindow(c *gin.Context)        { GetTimeWindowHandler(c) }

// ==================== System ====================
func GetSystemScale(c *gin.Context)     { GetSystemScaleHandler(c) }
func RefreshSystemScale(c *gin.Context) { RefreshSystemScaleHandler(c) }
func GetWarmupStatus(c *gin.Context)    { GetWarmupStatusHandler(c) }
func GetIndexes(c *gin.Context)         { GetIndexesHandler(c) }
func EnsureIndexes(c *gin.Context)      { EnsureIndexesHandler(c) }

// ==================== Storage ====================
func GetStorageConfig(c *gin.Context)    { GetStorageConfigHandler(c) }
func UpdateStorageConfig(c *gin.Context) { UpdateStorageConfigHandler(c) }
func CleanupCache(c *gin.Context)        { CleanupCacheHandler(c) }